% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GGUM MCMC Proposal Acceptance}
\alias{GGUM MCMC Proposal Acceptance}
\alias{acceptanceTheta}
\alias{acceptanceAlpha}
\alias{acceptanceDelta}
\alias{acceptanceTau}
\title{GGUM acceptance}
\usage{
acceptanceTheta(responses, cv, alphas, deltas, taus, SD)

acceptanceAlpha(responses, thetas, cv, delta, taus, SD)

acceptanceDelta(responses, thetas, alpha, cv, taus, SD)

acceptanceTau(k, responses, thetas, alpha, delta, taus, SD)
}
\arguments{
\item{responses}{A numeric vector giving the responses by the respondent
or item of interest; for \code{acceptanceTheta} this would be one row of
the response matrix, while for \code{acceptanceAlpha},
\code{acceptanceDelta}, or \code{acceptanceTau} this would be one column
of the response matrix}

\item{cv}{The current value of the paramenter of interest}

\item{alphas}{The numeric vector of current values of the items' alpha
parameters}

\item{deltas}{The numeric vector of current values of the items' delta
parameters}

\item{taus}{The list or vector of current values of the items' tau
parameters (for \code{acceptanceTheta} this is a list because there are
multiple tau vectors that must be used, but for the other acceptance
functions, only one numeric vector is needed)}

\item{SD}{The sigma parameter to be used for the proposal distribution}

\item{thetas}{The numeric vector of current values of the items' theta
parameters}

\item{delta}{The delta parameter for the item of interest (for
\code{acceptanceAlpha} and \code{acceptanceTau}, there is only one
delta needed for the calculations)}

\item{alpha}{The alpha parameter for the item of interest (for
\code{acceptanceDelta} and \code{acceptanceTau}, there is only one
alpha needed for the calculations)}

\item{k}{For \code{acceptanceTau}, an integer giving the index
(NOTE: the index must be specified as the vector would be indexed in
C++; i.e., the R index - 1 since C++ begins indexing at 0) in the
tau vector of the tau parameter of interest.}
}
\value{
If the acceptance ratio is greater than one or greater than a
  number randomly generated from the standard uniform distribution,
  the proposed value is returned; otherwise, the \code{cv} is returned.
}
\description{
Determines a new value for the parameters used in the MCMC algorithm
for the GGUM.
}
\details{
Given a current value \code{cv} for the parameter of interest,
as well as the responses and other parameters relevant to estimating the
parameter of interest, a new proposal for the parameter is generated
and accepted with probability
\deqn{\min \{1, \frac{\mathcal{L}(X|\theta^*)\pi(\theta^*)}{%
\mathcal{L}(X|\theta)\pi(\theta)}\}}{%
min {1, (L(X|\theta*)\pi(\theta*) / (L(X|\theta)\pi(\theta))}}
where \eqn{\theta^*}{\theta*} is the proposed value,
\eqn{\theta} is the current value, \eqn{\pi(\cdot)}{\pi(.)}
is the prior probability a parameter takes a value, and
\eqn{\mathcal{L}(X|\cdot)}{L(X|.)} is the likelihood of observing the
responses given a parameter value.
}
\section{Warning}{

For \code{acceptanceTau}, the tau vector is indexed beginning with 0!
When using this function, you must supply an index of one less than the
index you would supply if subsetting in \code{R}.
}

