% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Four Parameter Beta}
\alias{Four Parameter Beta}
\alias{d4beta}
\alias{p4beta}
\alias{q4beta}
\alias{r4beta}
\title{The Four Parameter Beta Distribution}
\usage{
d4beta(x, shape1, shape2, a, b)

p4beta(q, shape1, shape2, a, b)

q4beta(p, shape1, shape2, a, b)

r4beta(n, shape1, shape2, a, b)
}
\arguments{
\item{x, q}{A numeric vector of quantiles}

\item{shape1}{Shape 1}

\item{shape2}{Shape 2}

\item{a}{The lowest possible value of the distribution}

\item{b}{The highest possible value of the distribution}

\item{p}{A numeric vector of probabilities}

\item{n}{The number of observations to generate}
}
\value{
\code{d4beta} gives gives the density, \code{p4beta} gives the
  distribution function, \code{q4beta} gives the quantile function, and
  \code{r4beta} generates random numbers from the distribution.
}
\description{
Provides probability density, cumulative distribution, quantile, and random
number generating functions for the four parameter beta distribution.
}

