% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getPriorTheta}
\alias{getPriorTheta}
\title{GGUM getPrior}
\usage{
getPriorTheta(cv)
}
\arguments{
\item{cv}{A numeric vector of length one; the current value of the
parameter of interest}
}
\value{
A numeric vector of length one. The prior probability of observing
  \code{cv} for the paramenter of interest.
}
\description{
Get the prior probability of values for the GGUM parameters theta, alpha,
delta, and tau.
}
