% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglikelihoodRow}
\alias{loglikelihoodRow}
\alias{loglikelihoodCol}
\title{GGUM Log Likelihood}
\usage{
loglikelihoodRow(responses, theta, alphas, deltas, taus)

loglikelihoodCol(responses, thetas, alpha, delta, taus)
}
\arguments{
\item{responses}{For \code{loglikelihoodRow}, a numeric vector of length
length n (the number of items) giving the option chosen by the individual
for each item j; for \code{loglikelihoodCol}, a numeric vector of length
N (the number of respondents), giving the option chosen by each individual
i to the item}

\item{theta}{For \code{loglikelihoodRow}, a numeric vector of length one
giving the individual's latent trait parameter}

\item{alphas}{A numeric vector of length n; each element of the vector is an
item's discrimination parameter}

\item{deltas}{A numeric vector of length n; each element of the vector is an
item's location parameter}

\item{taus}{For \code{loglikelihoodRow}, a list of numeric vectors where
each list element j is a numeric vector of threshold parameters for item
j's options (where the first element of the vector should be zero);
for \code{loglikelihoodCol}, this is only the numeric vector of threshold
parameters for the item of interest}

\item{thetas}{A numeric vector of length N, each each element of which is
an individual's latent trait parameter}

\item{alpha}{A numeric vector of length one giving the item's
discrimination parameter}

\item{delta}{A numeric vector of length one giving the item's
location parameter}
}
\value{
The (log) likelihood of the vector of interest.
}
\description{
Calculate the log likelihood of data for the GGUM given parameter values.
}
\details{
This function calculates the log likelihood of a \textbf{vector} of
responses given values for the parameters relevant to the responses.
This could be all of a respondent \eqn{i}'s responses to every item \eqn{j},
or all respondents' responses to an item \eqn{j}. We calculate likelihood
of vectors rather than the likelihood of the entire response matrix since
when computing acceptance ratios for the MCMC algorithm, we divide products
of probabilities, most of which will cancel out -- thus, we only ever need
the likelihood of one vector at any given time.
}
