% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglikelihoodAlpha}
\alias{loglikelihoodAlpha}
\title{GGUM LogLikelihood Alpha, Delta, and Taus.}
\usage{
loglikelihoodAlpha(responses, thetas, alpha, delta, taus)
}
\arguments{
\item{thetas}{A numeric vector of length N (the number of respondents); each
each element of the vector is an individual's latent trait parameter}

\item{alpha}{A numeric vector of length one giving the item's discrimination
parameter}

\item{delta}{A numeric vector of length one giving the item's location
parameter}

\item{taus}{A numeric vector of length K (the number of options) giving the
threshold parameter for each option; the first element should be zero.}

\item{responseVector}{A numeric vector of length n (the number of items)
giving the option chosen by individual i for the item}
}
\value{
The (log) likelihood of the vector of interest.
}
\description{
Calculate the Log Likelihood parameters Alpha, Delta, and Taus
for the GGUM.
}
