% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglikelihoodTheta}
\alias{loglikelihoodTheta}
\title{GGUM LogLikelihood - Theta}
\usage{
loglikelihoodTheta(responses, theta, alphas, deltas, taus)
}
\arguments{
\item{theta}{A numeric vector of length one giving the individual's latent
trait parameter}

\item{alphas}{A numeric vector of length n; each element of the vector is an
item's discrimination parameter}

\item{deltas}{A numeric vector of length n; each element of the vector is an
item's location parameter}

\item{taus}{A list of numeric vectors; each list element j is a numeric
vector of threshold parameters for item j's options (where the first
element of the vector should be zero).}

\item{responseVector}{A numeric vector of length N (the number of
respondents) giving the option chosen the individual for each item j}
}
\value{
The (log) likelihood of the vector of interest.
}
\description{
Calculate the Log Likelihood parameters Theta for the GGUM.
}
