% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ggumMC3}
\alias{ggumMC3}
\title{GGUM MC3}
\usage{
ggumMC3(data, iters, r_one, r_one_val, r_two, r_two_val, N, W)
}
\arguments{
\item{data}{A numeric matrix giving the individuals' responses}

\item{iters}{A vector of length one giving the number of iterations}

\item{r_one}{The index of the first respondent whose theta is restricted}

\item{r_one_val}{The value of the first theta restriction}

\item{r_two}{The index of the second respondent whose theta is restricted}

\item{r_two_val}{The value of the second theta restriction}

\item{N}{The number of chains}

\item{W}{The period by which to attempt chain swaps; e.g. if W = 100,
a state swap will be proposed between two randomly selected chains
every 100 iterations}
}
\value{
A numeric matrix giving the parameter values at each iteration
  for the cold chain
}
\description{
Metropolis Coupled Markov Chain Monte Carlo Sampling for the GGUM
}
