% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ggumMCMC}
\alias{ggumMCMC}
\title{GGUM MCMC Sampler}
\usage{
ggumMCMC(responseMatrix, Kvector, iterations)
}
\arguments{
\item{responseMatrix}{A numeric matrix giving the response by each
respondent to each item}

\item{Kvector}{A numeric vector of length m (the number of items), each
element j of which gives the number of options (K) for item j}

\item{iterations}{A numeric vector of length one; the number of iterations}
}
\value{
A chain matrix; a numeric matrix with \code{iterations} rows and
  one column for every parameter of the model, so that each element of the
  matrix gives the value of a parameter for a particular iteration of the
  algorithm
}
\description{
MCMC sampler for the generalized graded unfolding model (GGUM), utilizing
a Metropolis-Hastings algorithm
}
