% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ggumProbability}
\alias{ggumProbability}
\title{GGUM Probability Function}
\usage{
ggumProbability(k, theta, alpha, delta, tau)
}
\arguments{
\item{k}{A numeric vector giving the item(s) for which response
probability should be calculated}

\item{theta}{A numeric vector of length one giving the respondent's latent
trait score}

\item{alpha}{A numeric vector of length one giving the item's
discrimination parameter}

\item{delta}{A numeric vector of length one giving the item's location
parameter}

\item{tau}{A numeric vector of length K (the number of possible responses)
giving the options' threshold parameters; the first element of \code{tau}
should be zero}
}
\value{
A numeric vector the same length of \code{k} giving the response
  probabilities
}
\description{
Calculate the probability of a response according to the GGUM
}
\details{
The General Graded Unfolding Model (GGUM) is an item response model
designed to consider the possibility of disagreement for opposite reasons.
This function gives the probability of a respondent's response to a test
item given item and respondent parameters. The user can calculate the
probability of one particular response or for any number of the possible
responses to the item (see the \code{k} parameter described below).

The probability that respondent \eqn{i} chooses option \eqn{k} for item
\eqn{j} is given by
\deqn{\frac{\exp (\alpha_j [k (\theta_i - \delta_j) -
\sum_{m=0}^k \tau_{jm}]) + \exp (\alpha_j [(2K - k - 1)
(\theta_i - \delta_j) - \sum_{m=0}^k \tau_{jm}])}{%
\sum_{l=0}^{K-1} [\exp (\alpha_j [l (\theta_i - \delta_j) -
\sum_{m=0}^l \tau_{jm}]) + \exp (\alpha_j [(2K - l - 1)
(\theta_i - \delta_j) - \sum_{m=0}^l \tau_{jm}])]}}{%
(exp(\alpha_j [k(\theta_i-\delta_j) - 
\Sigma_{m=0}^k \tau_{jm}]) + exp(\alpha_j [(2K - k - 1)
(\theta_i - \delta_j) - \Sigma_{m=0}^k \tau_{jm}])) /
(\Sigma_{l=0}^{K-1} [exp (\alpha_j [l (\theta_i - \delta_j) -
\Sigma_{m=0}^l \tau_{jm}]) + exp (\alpha_j [(2K - l - 1)
(\theta_i - \delta_j) - \Sigma_{m=0}^l \tau_{jm}])])},
where \eqn{\theta_i} is \eqn{i}'s latent trait parameter,
\eqn{\alpha_j} is the item's discrimination paramter,
\eqn{\delta_j} is the item's location paramter,
\eqn{\tau_{j0}, \ldots, \tau_{j(K-1)}} are the options' threshold
parameters, and \eqn{\tau_{j0}} is 0,
\eqn{K} is the number of options for item \eqn{j}, and
the options are indexed by \eqn{k = 0, \ldots, K-1}.
}
\references{
Roberts, James S., John R. Donoghue, and James E. Laughlin.
  2000. ``A General Item Response Theory Model for Unfolding
  Unidimensional Polytomous Responses." \emph{Applied Psychological
  Measurement} 24(1): 3--32.

de la Torre, Jimmy, Stephen Stark, and Oleksandr S.
  Chernyshenko. 2006. ``Markov Chain Monte Carlo Estimation of Item
  Parameters for the Generalized Graded Unfolding Model." \emph{Applied
  Psychological Measurement} 30(3): 216--232.
}
