% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GGUM MCMC Proposal Densities}
\alias{GGUM MCMC Proposal Densities}
\alias{proposerTheta}
\alias{proposerTheta,}
\alias{proposerAlpha,}
\alias{proposerDelta,}
\alias{proposerTau,}
\alias{proposer}
\alias{proposerAlpha}
\alias{proposerDelta}
\alias{proposerTau}
\title{GGUM Proposer}
\usage{
proposerTheta(cv, SD)

proposerAlpha(cv, SD)

proposerDelta(cv, SD)

proposerTau(cv, SD)
}
\arguments{
\item{cv}{A numeric vector of length one;
the current value of the parameter of interest.}

\item{SD}{A numeric vector of length one;
the sigma parameter used to generate the proposed value.}
}
\value{
A numeric vector of length one;
  the proposed value for the parameter of interest.
}
\description{
Propose new values for the GGUM parameters Theta, Alpha, Delta, and Tau
}
\details{
Given a current value of the parameter of interest,
and a sigma parameter for the proposal distribution,
a new proposal for the parameter's value is given.
The following proposal densities are used:
\itemize{
  \item Theta -- A location-scale T distribution shifted by \code{cv} and
    scaled by \code{SD}, truncated at -10 and 10
  \item Alpha -- A location-scale T distribution shifted by \code{cv} and
    scaled by \code{SD}, truncated at 0.25 and 4
  \item Delta -- A location-scale T distribution shifted by \code{cv} and
    scaled by \code{SD}, truncated at -5 and 5
  \item Tau -- A location-scale T distribution shifted by \code{cv} and
    scaled by \code{SD}, truncated at -6 and 6
}
}
