% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggumR-package.R
\docType{package}
\name{ggumR}
\alias{ggumR}
\alias{ggumR-package}
\title{ggumR}
\description{
ggumR provides R tools for the generalized graded unfolding model.
}
\author{
JB Duck-Mayr, Patrick Cunha Silva, Luwei Ying
}
