% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterationsTheta.R
\name{interationsTheta}
\alias{interationsTheta}
\title{GGUM interations Only Thetas}
\usage{
interationsTheta(startvalue, iterations, responseMatrix, thetas, alphas, deltas,
  taus)
}
\arguments{
\item{startvalue}{A numeric vector of length 1.}

\item{iterations}{A numeric vector of length 1.}

\item{responseMatrix}{A numeric matrix with N rows and n (the number of
items) columns; each i,j element of the matrix gives the option chosen by
individual i for item j}

\item{thetas}{A numeric vector of length N (the number of respondents); each
each element of the vector is an individual's latent trait parameter}

\item{alphas}{A numeric vector of length n; each element of the vector is an
item's discrimination parameter}

\item{deltas}{A numeric vector of length n; each element of the vector is an
item's location parameter}

\item{taus}{A list of numeric vectors; each list element j is a numeric
vector of threshold parameters for item j's options (where the first
element of the vector should be zero).}
}
\description{
\code{MCMC} runs the MCMC algorithm  for GGUM
}

