% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterateAlphas.R
\name{iterateAlphas}
\alias{iterateAlphas}
\title{GGUM iterations for Alphas}
\usage{
iterateAlphas(thetas, deltas, taus, times, responseMatrix, K)
}
\arguments{
\item{thetas}{A numeric vector of length N; each element of the vector is a
respondent's latent trait parameter}

\item{deltas}{A numeric vector of length n; each element of the vector is an
item's location parameter}

\item{taus}{A list of numeric vectors; each list element j is a numeric
vector of threshold parameters for item j's options (where the first
element of the vector should be zero).}

\item{times}{A numeric vector of length 1.}

\item{responseMatrix}{A numeric matrix with N rows and n (the number of
items) columns; each i,j element of the matrix gives the option chosen by
individual i for item j}

\item{K}{A numeric vector of length one giving the number of options for
each item}
}
\value{
A numeric matrix where each i,j entry is the value for alpha j at
  iteration i
}
\description{
\code{iterateAlphas} runs the GGUM MCMC algorithm for alphas
}
