% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterateTaus.R
\name{iterateTaus}
\alias{iterateTaus}
\title{GGUM iterations for Taus}
\usage{
iterateTaus(thetas, alphas, deltas, times, responseMatrix, K)
}
\arguments{
\item{thetas}{A numeric vector of length N; each element of the vector is a
respondent's latent trait parameter}

\item{alphas}{A numeric vector of length n; each element of the vector is an
item's discrimination parameter}

\item{deltas}{A numeric vector of length n; each element of the vector is an
item's location parameter}

\item{times}{A numeric vector of length 1.}

\item{responseMatrix}{A numeric matrix with N rows and n (the number of
items) columns; each i,j element of the matrix gives the option chosen by
individual i for item j}

\item{K}{A numeric vector of length one giving the number of options for
each item}
}
\value{
A numeric matrix where each i,j entry is the value for
  tau[[((j - 1) %/% K) + 1]][((j - 1) %% K) + 1] at iteration i
}
\description{
\code{iterateTaus} runs the GGUM MCMC algorithm for taus
}
