% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterateThetas.R
\name{iterateThetas}
\alias{iterateThetas}
\title{GGUM iterations for Thetas}
\usage{
iterateThetas(alphas, deltas, taus, times, responseMatrix, K)
}
\arguments{
\item{alphas}{A numeric vector of length n; each element of the vector is an
item's discrimination parameter}

\item{deltas}{A numeric vector of length n; each element of the vector is an
item's location parameter}

\item{taus}{A list of numeric vectors; each list element j is a numeric
vector of threshold parameters for item j's options (where the first
element of the vector should be zero).}

\item{times}{A numeric vector of length 1 giving the number of iterations
the algorithm should run for.}

\item{responseMatrix}{A numeric matrix with N rows and n (the number of
items) columns; each i,j element of the matrix gives the option chosen by
individual i for item j}

\item{K}{A numeric vector of length one giving the number of options for
each item}
}
\value{
A numeric matrix where each i,j entry is the value for theta j at
  iteration i
}
\description{
\code{iterateThetas} runs the GGUM MCMC algorithm for thetas
}
