% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodCol.R
\name{likelihoodCol}
\alias{likelihoodCol}
\title{GGUM Likelihood By Columns}
\usage{
likelihoodCol(thetas, responseMatrix, alphas, deltas, taus, index = 1)
}
\arguments{
\item{thetas}{A numeric vector of length N (the number of respondents); each
each element of the vector is an individual's latent trait parameter}

\item{responseMatrix}{A numeric matrix with N rows and n (the number of
items) columns; each i,j element of the matrix gives the option chosen by
individual i for item j}

\item{alphas}{A numeric vector of length n; each element of the vector is an
item's discrimination parameter}

\item{deltas}{A numeric vector of length n; each element of the vector is an
item's location parameter}

\item{taus}{A list of numeric vectors; each list element j is a numeric
vector of threshold parameters for item j's options (where the first
element of the vector should be zero).}

\item{index}{A numeric vector of length one giving the row/column to be
included in the calculation}
}
\value{
The likelihood of the matrix or vector of interest.
}
\description{
\code{likelihoodCol} gives the likelihood  of the paramenters (Columns)  of 
a response matrix or vector thereof
}
\details{
Given parameters for the individuals and test items, a response matrix
recording each individual's response to each test item, and as well as
(optionally) an indication of the vector of interest, \code{likelihood}
gives the product of the probability corresponding to each entry in the
matrix or relevant vector.
}
