% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llCol.R
\name{llCol}
\alias{llCol}
\title{GGUM Likelihood By Columns}
\usage{
llCol(thetas, responseVector, alpha, delta, taus)
}
\arguments{
\item{thetas}{A numeric vector of length N (the number of respondents); each
each element of the vector is an individual's latent trait parameter}

\item{responseVector}{A numeric vector of length n (the number of items)
giving the option chosen by individual i for the item}

\item{alpha}{A numeric vector of length one giving the item's discrimination
parameter}

\item{delta}{A numeric vector of length one giving the item's location
parameter}

\item{taus}{A numeric vector of length K (the number of options) giving the
threshold parameter for each option; the first element should be zero.}
}
\value{
The (log) likelihood of the vector of interest.
}
\description{
\code{llCol} gives the (log) likelihood for one column of a response matrix
}
\details{
Given parameters for the individuals and test items, a response vector
recording each individual's response to the test item, \code{llCol} gives
the product of the probability corresponding to each entry in the relevant
vector.
}
