% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llRow.R
\name{llRow}
\alias{llRow}
\title{GGUM Likelihood By Rows}
\usage{
llRow(theta, responseVector, alphas, deltas, taus)
}
\arguments{
\item{theta}{A numeric vector of length one giving the individual's latent
trait parameter}

\item{responseVector}{A numeric vector of length N (the number of
respondents) giving the option chosen the individual for each item j}

\item{alphas}{A numeric vector of length n; each element of the vector is an
item's discrimination parameter}

\item{deltas}{A numeric vector of length n; each element of the vector is an
item's location parameter}

\item{taus}{A list of numeric vectors; each list element j is a numeric
vector of threshold parameters for item j's options (where the first
element of the vector should be zero).}
}
\value{
The (log) likelihood of the vector of interest.
}
\description{
\code{llRow} gives the (log) likelihood for one row of a response matrix
}
\details{
Given parameters for the individuals and test items, a response vector
recording an individual's response to each test item, \code{llRow} gives
the product of the probability corresponding to each entry in the relevant
vector.
}
