% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{probability}
\alias{probability}
\title{GGUM Probability Function}
\usage{
probability(k, K, theta, alpha, delta, tau)
}
\arguments{
\item{k}{A numeric vector; the index of the item option of interest}

\item{K}{A numeric vector of length one; the number of options for the item}

\item{theta}{A numeric vector of length one; the individual's latent trait
parameter}

\item{alpha}{A numeric vector of length one; the item's discrimination
parameter}

\item{delta}{A numeric vector of length one; the item's location parameter}

\item{tau}{A numeric vector of length K; the vector threshold parameters
for the item's options (where the first element of tau should be zero).}
}
\value{
The probability that resopndent will choose response category k.
}
\description{
\code{probability} gives the probability a person will choose an item option.
}
\details{
Given parameters for the individual and the test item, \code{probabiliy}
returns the probability the individual will choose a particular response
option for the test item according to the GGUM.
}

