% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probMatrix.R
\name{probMatrix}
\alias{probMatrix}
\title{GGUM Probability of Response Matrix}
\usage{
probMatrix(thetas, responseMatrix, alphas, deltas, taus)
}
\arguments{
\item{thetas}{A numeric vector of length N (the number of respondents); each
each element of the vector is an individual's latent trait parameter}

\item{responseMatrix}{A numeric matrix with N rows and n (the number of
items) columns; each i,j element of the matrix gives the option chosen by
individual i for item j}

\item{alphas}{A numeric vector of length n; each element of the vector is an
item's discrimination parameter}

\item{deltas}{A numeric vector of length n; each element of the vector is an
item's location parameter}

\item{taus}{A list of numeric vectors; each list element j is a numeric
vector of threshold parameters for item j's options (where the first
element of the vector should be zero).}
}
\value{
A matrix probabilities for the elements of the response matrix.
}
\description{
\code{probMatrix} gives the probability of each observed response
}
\details{
Given parameters for the individuals and test items, and a response matrix
recording each individual's response to each test item, \code{probMatrix}
returns a matrix of probabilities; the i,j entry in the matrix is the
probability individual i chose the particular response to item j that we
observed for individual i and item j (i.e. the probability corresponding to
the i,j entry in the response matrix).
}
