% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposer.R
\name{proposerTheta}
\alias{proposerTheta}
\title{Helper function for \code{GGUM}}
\usage{
proposerTheta(currentValue, sdPeriodType = 1)
}
\arguments{
\item{currentValue}{The current value of the parameter of interest}

\item{sdPeriodType}{The standard-deviation used in the t-scaled distribution
to draw the proposer.}
}
\value{
A proposed value for the paramenters used in the MCMC algorithm for the GGUM.
}
\description{
\code{proposer} generates proposed values for parameters.
}

