% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler}
\alias{sampler}
\title{Function for \code{GGUM}}
\usage{
sampler(responseMatrix, times = 7000, thetas, alphas, deltas, taus,
  startTheta = NULL, startAlpha = NULL, startDelta = NULL)
}
\arguments{
\item{responseMatrix}{A numeric matrix with N (the number of individuals)
rows and n (the number of items) columns; each i,j element of the matrix
gives the option chosen by individual i for item j}

\item{thetas}{The vector of current values of the items' delta parameters}

\item{alphas}{The vector of current values of the items' alpha parameters}

\item{deltas}{The vector of current values of the items' delta parameters}

\item{taus}{The list of current values of the items' tau vectors}
}
\value{
the iteration matrix with K (the number of iteration stages) rows and n (the number of items) columns.
}
\description{
\code{sampler} takes in a matrix of all responses and returns the iteration matrix.
}

