% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Location-Scale T Distribution}
\alias{Location-Scale T Distribution}
\alias{dScaledT}
\alias{pScaledT}
\alias{qScaledT}
\alias{rScaledT}
\alias{r_TruncNorm}
\title{The location-scale Student's T distribution.}
\usage{
dScaledT(x, df, mu, sigma)

pScaledT(q, df, mu, sigma)

qScaledT(p, df, mu, sigma)

rScaledT(n, df, mu, sigma)

r_TruncNorm(mu, sigma, a, b)
}
\arguments{
\item{x, q}{A numeric vector of the quantiles of interest}

\item{df}{A numeric vector of length one; the degrees of freedom of the
distribution}

\item{mu}{A numeric vector of length one; the shifting parameter}

\item{sigma}{A numeric vector of length one; the scale parameter}

\item{p}{A numeric vector of the probabilities of interest}

\item{n}{A numeric vector of length one; the number of numbers to generate}
}
\value{
'dScaledT' gives the density, 'pScaledT' gives the distribution
    function, 'qScaledT' gives the quantile function, and 'rScaledT'
    generates random numbers from the distribution
}
\description{
Density, distribution function, quantile function, and random number
generation for the Student's T distribution shifted by location parameter
'mu' and scaled by 'sigma'.
}
\references{
Jackman, Simon. 2009. \emph{Bayesian Analysis for the Social
    Sciences}. Wiley. p. 507.
}

