% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stableProd.R
\name{stableProd}
\alias{stableProd}
\title{Numerically stable product}
\usage{
stableProd(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The product of the elements of x.
}
\description{
\code{stableProd} calculates the product of a numeric vector
}
\details{
While base R provides the function \code{prod}, a more numerically stable
way to calculate the product of a vector is to take the exponential of the
sum of the natural logarithm of each element of the vector.
}
