% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Truncated Location-Scale T}
\alias{Truncated Location-Scale T}
\alias{dtrunclst}
\alias{ptrunclst}
\alias{qtrunclst}
\alias{rtrunclst}
\title{The Truncated Location-Scale T distribution.}
\usage{
dtrunclst(x, df, mu, sigma, a, b)

ptrunclst(q, df, mu, sigma, a, b)

qtrunclst(p, df, mu, sigma, a, b)

rtrunclst(n, df, mu, sigma, a, b)
}
\arguments{
\item{x, q}{A numeric vector of quantiles}

\item{df}{A numeric vector of length one; the degrees of freedom of the
distribution}

\item{mu}{A numeric vector of length one; the shifting parameter}

\item{sigma}{A numeric vector of length one; the scale parameter}

\item{a}{The lowest possible value of the distribution}

\item{b}{The highest possible value of the distribution}

\item{p}{A numeric vector of probabilities}

\item{n}{The number of observations to generate}
}
\value{
\code{dtrunclst} gives gives the density, \code{ptrunclst} gives
  the distribution function, \code{qtrunclst} gives the quantile function,
  and \code{rtrunclst} generates random numbers from the distribution.
}
\description{
Provides probability density, cumulative distribution, quantile, and random
number generating functions for the truncated location-scale T distribution.
}
