% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Truncated Normal}
\alias{Truncated Normal}
\alias{dtruncnorm}
\alias{ptruncnorm}
\alias{qtruncnorm}
\alias{rtruncnorm}
\title{The Truncated Normal Distribution}
\usage{
dtruncnorm(x, mean, SD, a, b)

ptruncnorm(q, mean, SD, a, b)

qtruncnorm(p, mean, SD, a, b)

rtruncnorm(n, mean, SD, a, b)
}
\arguments{
\item{x, q}{A numeric vector of quantiles}

\item{mean}{The mean of the distribution}

\item{SD}{The standard deviation of the distribution}

\item{a}{The lowest possible value of the distribution}

\item{b}{The highest possible value of the distribution}

\item{p}{A numeric vector of probabilities}

\item{n}{The number of observations to generate}
}
\value{
\code{dtruncnorm} gives gives the density, \code{ptruncnorm} gives
  the distribution function, \code{qtruncnorm} gives the quantile function,
  and \code{rtruncnorm} generates random numbers from the distribution.
}
\description{
Provides probability density, cumulative distribution, quantile, and random
number generating functions for the truncated normal distribution.
}
