% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptance.R
\name{acceptanceTheta}
\alias{acceptanceTheta}
\title{Helper function for \code{GGUM}}
\usage{
acceptanceTheta(currentValue, proposedValue, responseMatrix, currentValuesTheta,
  proposedValuesTheta, alphas, deltas, taus, index = 1)
}
\arguments{
\item{currentValue}{The current value of the paramenter of interest}

\item{proposedValue}{The proposed value for the paramenter of interest}

\item{responseMatrix}{A numeric matrix with N (the number of individuals)
rows and n (the number of items) columns; each i,j element of the matrix
gives the option chosen by individual i for item j}

\item{alphas}{The vector of current values of the items' alpha parameters}

\item{deltas}{The vector of current values of the items' delta parameters}

\item{taus}{The list of current values of the items' tau vectors}

\item{thetas}{The vector of current values of the items' delta parameters}
}
\value{
The acceptance ratio for the paramenters used in the MCMC algorithm for the GGUM.
}
\description{
\code{acceptanceTheta} computes a acceptance ratio for the paramenters used 
in the MCMC algorithm for the GGUM.
}

