% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{acceptanceTheta}
\alias{acceptanceTheta}
\title{GGUM acceptance}
\usage{
acceptanceTheta(responses, cv, alphas, deltas, taus, SD)
}
\arguments{
\item{responses}{A numeric vector giving the response by each respondent to each item}

\item{cv}{The current value of the paramenter of interest}

\item{alphas}{The numeric vector of current values of the items' alpha parameters}

\item{SD}{The standard deviation of the truncated location-scale T distribution}

\item{thetas}{The numeric vector of current values of the items' theta parameters}

\item{delta}{The numeric vector of current values of the items' delta parameters}

\item{thetas}{The vector of current values of the items' delta parameters}
}
\value{
The proposed value or current value, according to the acceptance ratio, to
be used in the MCMC algorithm for the GGUM.
}
\description{
Computes an acceptance ratio for the paramenters used in the MCMC algorithm
for the GGUM.
}

