% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrior.R
\name{getPrior}
\alias{getPrior}
\title{Get priors for GGUM parameters}
\usage{
getPrior(param, value)
}
\arguments{
\item{param}{A character vector indicating the parameter type; should be one
of 'alpha', 'delta', 'tau', or 'theta'}

\item{value}{The parameter's value}
}
\value{
The probability of drawing value from a prior distribution for param.
}
\description{
\code{getPrior} gives the probability of drawing value from the prior
distribution for param.
}
