% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GGUM Priors}
\alias{GGUM Priors}
\alias{getPriorTheta}
\alias{getPriorAlpha}
\alias{getPriorDelta}
\alias{getPriorTaus}
\title{GGUM getPrior}
\usage{
getPriorTheta(cv)

getPriorAlpha(cv)

getPriorDelta(cv)

getPriorTaus(cv)
}
\arguments{
\item{cv}{A numeric vector of length one; the current value of the
parameter of interest}
}
\value{
A numeric vector of length one. The prior probability of observing
  \code{cv} for the paramenter of interest.
}
\description{
Get the prior probability of values for the GGUM parameters theta, alpha,
delta, and tau.
}
\details{
Following de la Torre et al (2006), we use the following prior
distributions for MCMC estimation of GGUM parameters:
\itemize{
  \item Theta -- The standard normal
  \item Alpha -- A four parameter beta distribution, with shape parameters
    1.5, 1.5, minimum value 0.25, and maximum value 4
  \item Delta -- A four parameter beta distribution, with shape parameters
    2, 2, minimum value -5, and maximum value 5
  \item Tau -- A four parameter beta distribution, with shape parameters
    2, 2, minimum value -6, and maximum value 6
}
}
\references{
de la Torre, Jimmy, Stephen Stark, and Oleksandr S.
  Chernyshenko. 2006. ``Markov Chain Monte Carlo Estimation of Item
  Parameters for the Generalized Graded Unfolding Model." \emph{Applied
  Psychological Measurement} 30(3): 216--232.
}
