% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ggumProbability}
\alias{ggumProbability}
\title{GGUM Probability Function}
\usage{
ggumProbability(k, theta, alpha, delta, tau)
}
\arguments{
\item{k}{A numeric vector giving the item(s) for which response
probability should be calculated}

\item{theta}{A numeric vector of length one giving the respondent's latent
trait score}

\item{alpha}{A numeric vector of length one giving the item's
discrimination parameter}

\item{delta}{A numeric vector of length one giving the item's location
parameter}

\item{tau}{A numeric vector of length K (the number of possible responses)
giving the options' threshold parameters; the first element of \code{tau}
should be zero}
}
\value{
A numeric vector the same length of \code{k} giving the response
  probabilities
}
\description{
Calculate the probability of a response according to the GGUM
}
\details{
The General Graded Unfolding Model (GGUM) is an item response model
designed to consider the possibility of disagreement for opposite reasons.
This function gives the probability of a respondent's response to a test
item given item and respondent parameters. The user can calculate the
probability of one particular response or for any number of the possible
responses to the item (see the \code{k} parameter described below).
}
