% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{proposerTheta}
\alias{proposerTheta}
\title{GGUM Proposer}
\usage{
proposerTheta(cv, SD)
}
\arguments{
\item{cv}{A numeric vector of length one. The current value of the parameter of interest.}

\item{SD}{A numeric vector of length one. The standard deviation used to generate the 
proposed value.}
}
\value{
A numeric vector of length one. The propose value for the paramenter of interest.
}
\description{
Propose new values for the parameters Theta, Alpha, Delta, and Taus
for the GGUM.
}

