% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Location-Scale T}
\alias{Location-Scale T}
\alias{dlst}
\alias{plst}
\alias{qlst}
\alias{rlst}
\title{The location-scale T distribution.}
\usage{
dlst(x, df, mu, sigma)

plst(q, df, mu, sigma)

qlst(p, df, mu, sigma)

rlst(n, df, mu, sigma)
}
\arguments{
\item{x, q}{A numeric vector of the quantiles of interest}

\item{df}{A numeric vector of length one; the degrees of freedom of the
distribution}

\item{mu}{A numeric vector of length one; the shifting parameter}

\item{sigma}{A numeric vector of length one; the scale parameter}

\item{p}{A numeric vector of the probabilities of interest}

\item{n}{A numeric vector of length one; the number of numbers to generate}
}
\value{
'dlst' gives the density, 'plst' gives the distribution function,
    'qlst' gives the quantile function, and 'rlst' generates random numbers
    from the distribution
}
\description{
Density, distribution function, quantile function, and random number
generation for the T distribution shifted by location parameter 'mu' and
scaled by 'sigma'.
}
\references{
Jackman, Simon. 2009. \emph{Bayesian Analysis for the Social
    Sciences}. Wiley. p. 507.
}

